Moved DownsampleTransform and SuperResolutionDataset classes from modules.py to dataset.py. This change is necessary as modules.py is intended for containing the source code of the components of the model. Data preprocessing classes will be stored in dataset.py as they are only used in that file.
